+++ src/models/model_const.h
@@ -4,9 +4,9 @@ class TModelConst : public TModel
 {
 public:
 	TModelConst();
-    virtual void Encode(uint8 symbol, uint32 &low_count, uint32 &upper_count, uint32 &total);
-    virtual uint8 Decode(uint32 value, uint32 &lower_count, uint32 &upper_count);
-    virtual uint32 GetNormalizer() {
+    virtual void Encode(uint8 symbol, uint64 &low_count, uint64 &upper_count, uint64 &total);
+    virtual uint8 Decode(uint64 value, uint64 &lower_count, uint64 &upper_count);
+    virtual uint64 GetNormalizer() {
         return Normalizer;
     };
     virtual void Observe(uint8 symbol);
@@ -16,6 +16,6 @@ public:
     };
 protected:
 
-	uint32 Probs[256];
-	uint32 Normalizer;
+	uint64 Probs[256];
+	uint64 Normalizer;
 };
